SLM-73: Moved prisonTypes to @types
