+++ server/@types/prisonTypes.ts
@@ -1,4 +1,4 @@
-export interface Prison {
+export type Prison = {
   id: string
   name: string
 }
